requirements.txt
fastapi
uvicorn
langchain
langchain_core
langchain_groq
langchain_community
pandas
pydantic
python-dotenv
Requests
SQLAlchemy
